Update wp-social-login.php
This update is to fix the error when trying to logging with Facebook. The error is:

"Invalid Scopes: user_friends. This message is only shown to developers. Users of your app will ignore these permissions if present. Please read the documentation for valid permissions at https://developers.facebook.com/docs/f…ermissions"

This fix was proposed by Miled here: https://wordpress.org/support/topic/error-when-logging-with-fb/#post-10347261